Redesign Grader, add Subject and Record
Switched from storing grades, criteria names, and weights
in separate arrays to creating a Record class that stores
information about grades, criteria names, and weights.
Redesign Grader to extend an ArrayList containing Records.
This allows one array list to store all of our relevant
data and for methods to act on data contained in the
array list. There are 3 errors and 10 warnings.